# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://www.swagger.io/support",
            "email": "support@swagger.io"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        }
    },
    "paths": {
        "/api/v1/participant": {
            "get": {
                "description": "Get a list of Participant.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Participant"
                ],
                "summary": "Get all Participant.",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/formatters.ParticipantFormatter"
                            }
                        }
                    }
                }
            },
            "post": {
                "description": "Creating a new Participant.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Participant"
                ],
                "summary": "Create New Participant.",
                "parameters": [
                    {
                        "description": "the body to create a new participant",
                        "name": "Body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/inputs.CreateParticipantInput"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/helper.Response"
                        }
                    }
                }
            }
        },
        "/api/v1/participant/certificate/{id}": {
            "get": {
                "description": "Print Name Tag a participant by id.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Participant"
                ],
                "summary": "Print Certificate.",
                "parameters": [
                    {
                        "type": "string",
                        "description": "participant id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/api/v1/participant/name-tag/{id}": {
            "get": {
                "description": "Print Name Tag a participant by id.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Participant"
                ],
                "summary": "Print Name Tag.",
                "parameters": [
                    {
                        "type": "string",
                        "description": "participant id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/api/v1/participant/{id}": {
            "get": {
                "description": "Get a Participant by id.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Participant"
                ],
                "summary": "Get Participant.",
                "parameters": [
                    {
                        "type": "string",
                        "description": "participant id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/formatters.ParticipantFormatter"
                        }
                    }
                }
            },
            "put": {
                "description": "Update participant by id.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Participant"
                ],
                "summary": "Update Participant.",
                "parameters": [
                    {
                        "type": "string",
                        "description": "participant id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "the body to update an participant",
                        "name": "Body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/inputs.UpdateParticipantInput"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/helper.Response"
                        }
                    }
                }
            },
            "delete": {
                "description": "Delete a participant by id.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Participant"
                ],
                "summary": "Delete one Participant.",
                "parameters": [
                    {
                        "type": "string",
                        "description": "participant id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/helper.Response"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "formatters.ParticipantFormatter": {
            "type": "object",
            "properties": {
                "business_name": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "full_name": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "is_print_certificate": {
                    "type": "boolean"
                },
                "is_print_name_tag": {
                    "type": "boolean"
                },
                "phone_number": {
                    "type": "string"
                }
            }
        },
        "helper.Meta": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "message": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                }
            }
        },
        "helper.Response": {
            "type": "object",
            "properties": {
                "data": {},
                "meta": {
                    "$ref": "#/definitions/helper.Meta"
                }
            }
        },
        "inputs.CreateParticipantInput": {
            "type": "object",
            "required": [
                "business_name",
                "email",
                "full_name",
                "phone_number"
            ],
            "properties": {
                "business_name": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "full_name": {
                    "type": "string"
                },
                "phone_number": {
                    "type": "string"
                }
            }
        },
        "inputs.UpdateParticipantInput": {
            "type": "object",
            "required": [
                "business_name",
                "email",
                "full_name",
                "id",
                "phone_number"
            ],
            "properties": {
                "business_name": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "full_name": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "phone_number": {
                    "type": "string"
                }
            }
        }
    }
}